pyproject.toml
[project]
name = "careerpulse"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "anthropic>=0.42.0",
    "openai>=1.0.0",
    "apscheduler>=3.10.0",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.7.0",
    "pymupdf>=1.24.0",
    "tenacity>=8.2.0",
    "aiosmtplib>=3.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "httpx>=0.28.0",
]
playwright = [
    "playwright>=1.40.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]
